requirements.txt
click==8.1.7
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pytest==7.4.4
requests==2.31.0
rich==13.7.0  # For better CLI output formatting
redis==5.0.1
celery==5.3.6  # For task queue
flower==2.0.1  # For monitoring Celery tasks 